Support bulk writes (updates and saves) (#10)
* Aligned types with server.

* Added support for bulk save.

* Implement support for bulkUpdate

+++ src/main/java/io/serialized/client/aggregate/cache/VersionedState.java
@@ -3,9 +3,9 @@ package io.serialized.client.aggregate.cache;
 public class VersionedState<T> {
 
   private final T state;
-  private final long version;
+  private final int version;
 
-  public VersionedState(T state, long version) {
+  public VersionedState(T state, int version) {
     this.state = state;
     this.version = version;
   }
@@ -14,7 +14,7 @@ public class VersionedState<T> {
     return state;
   }
 
-  public long version() {
+  public int version() {
     return version;
   }
